Directly use WebSMS API, send ID
Directly use the websms API instead of email to SMS conversion. Gives us more control. Importantly we can set the message sender number. This is important so people can add it to their contacts / bypass do not disturb / set a loud ringtone for that contact etc.

There is probably a clever way to combine send_sms and send_phonecall, and allow for flexibility for future services. In the end they are all post requests that take some set of parameters stored in the database and send the message with a specific key, getting back a response code. I started to go in this direction with the generic "postparameters" database entry, but more work is needed (variable message formatting, response code interpretation) and there are more pressing things